Refine code

+++ handler.py
@@ -4,10 +4,14 @@ import os
 import shutil
 import sys
 
-from pyarmor.pyarmor import (main as call_pyarmor, pytransform_bootstrap,
+from pyarmor.pyarmor import (main as pyarmor_main, pytransform_bootstrap,
                              get_registration_code, query_keyinfo,
                              version as pyarmor_version)
 from pyarmor.project import Project
+
+
+def call_pyarmor(args):
+    pyarmor_main(args)
 
 
 class BaseHandler():
@@ -34,7 +38,7 @@ class RootHandler(BaseHandler):
 
     def __init__(self, config):
         super().__init__(config)
-        self.children.extend([ProjectHandler, LicenseHandler])
+        self.children.extend([ProjectHandler(config), LicenseHandler(config)])
 
     def do_version(self, args=None):
         pytransform_bootstrap()
@@ -64,10 +68,43 @@ class RootHandler(BaseHandler):
 class ProjectHandler(BaseHandler):
 
     data_file = 'index.json'
+    temp_id = 0
 
     def __init__(self, config):
         super().__init__(config)
         self.name = 'project'
+
+    def do_build_temp(self, args):
+        name = 'project-%s' % self.temp_id
+        path = os.path.join(self._get_path(), name)
+
+        if os.path.exists(path):
+            shutil.rmtree(path)
+        os.mkdir(path)
+
+        cmd_args = ['init', '--src', path, path]
+        call_pyarmor(cmd_args)
+
+        data = args.get('config', {})
+        project = Project()
+        project.open(path)
+
+        project._update(data)
+        project.save(path)
+
+        cmd_args = ['build']
+        target = args.get('traget')
+        if target and target not in ('obfuscate', 'default'):
+            cmd_args.extend(['--target', target])
+        output = args.get('output')
+        if output:
+            cmd_args.extend(['--output', output])
+
+        cmd_args.append(path)
+        call_pyarmor(cmd_args)
+
+        return output if output else os.path.join(path, 'dist')
+
 
     def do_new(self, args):
         c = self._get_config()
@@ -88,6 +125,7 @@ class ProjectHandler(BaseHandler):
         project = {
             'name': name,
             'title': args.get('title', name),
+            'target': args.get('target', ''),
             'path': os.path.abspath(path)
         }
 
@@ -102,29 +140,43 @@ class ProjectHandler(BaseHandler):
 
     def do_update(self, args):
         c, p = self._get_project(args)
+        if 'target' in args:
+            p['target'] = args['target']
+            self._set_config(c)
 
-        for x in ('name', 'title'):
-            if x in args:
-                p[x] = args[x]
+        data = args.get('config')
+        if data:
+            if 'name' in data:
+                p['name'] = data['name']
+                self._set_config(c)
+            if 'title' in data:
+                p['title'] = data['title']
                 self._set_config(c)
 
-        path = p['path']
-        project = Project()
-        project.open(path)
+            path = p['path']
+            project = Project()
+            project.open(path)
 
-        project._update(args)
-        project.save(path)
+            project._update(data)
+            project.save(path)
 
-        return project
+        return {
+            'id': args['id'],
+            'project': p,
+            'config': project
+        }
 
     def do_list(self, args):
         c = self._get_config()
-        return c['projects']
+        return c['projects'].items()
 
     def do_remove(self, args):
         c, p = self._get_project(args)
         c['projects'].remove(p)
         self._set_config(c)
+
+        if args.get('clean') and p['path'].startswith(self._get_path()):
+            shutil.rmtree(p['path'])
         return p
 
     def do_info(self, args):
@@ -139,13 +191,16 @@ class ProjectHandler(BaseHandler):
         path = p['path']
 
         cmd_args = ['build']
-        for k, v in args.items():
-            cmd_args.append('--' + k.replace('_', '-'))
-            if v is not None:
-                cmd_args.append(v)
+        target = args.get('traget')
+        if target and target not in ('obfuscate', 'default'):
+            cmd_args.extend(['--target', target])
+        output = args.get('output')
+        if output:
+            cmd_args.extend(['--output', output])
         cmd_args.append(path)
         call_pyarmor(cmd_args)
-        return True
+
+        return output if output else os.path.join(path, 'dist')
 
     def _get_project(self, args):
         c = self._get_config()
@@ -179,27 +234,36 @@ class ProjectHandler(BaseHandler):
 
 class LicenseHandler(BaseHandler):
 
+    counter_file = 'COUNTER'
+
     def __init__(self, config):
         super().__init__(config)
         self.name = 'license'
 
     def do_new(self, args):
         path = self._get_path()
-        cmd_args = ['licenses', '--output', path]
-        for k, v in args.items():
-            cmd_args.append('--' + k.replace('_', '-'))
-            if v is not None:
-                cmd_args.append(v)
+        output = args.get('output', path)
+        cmd_args = ['licenses', '--output', output]
+        cmd_args.extend(args.get('options', []))
+
+        rcode = args.get('rcode')
+        if rcode is None:
+            rcode = self._fetch_code()
+        cmd_args.append(rcode)
         call_pyarmor(cmd_args)
-        return path
+
+        return os.path.join(path, rcode, 'license.lic')
 
     def do_remove(self, args):
         path = self._get_path()
         rcode = args.get('rcode')
+        if not rcode:
+            raise RuntimeError('No registration code')
+
         licpath = os.path.join(path, rcode)
         if os.path.exists(licpath):
             shutil.rmtree(licpath)
-        return licpath
+        return os.path.join(licpath, 'license.lic')
 
     def do_list(self, args=None):
         path = self._get_path()
@@ -208,15 +272,33 @@ class LicenseHandler(BaseHandler):
     def do_info(self, args):
         path = self._get_path()
         rcode = args.get('rcode')
+        if not rcode:
+            raise RuntimeError('No registration code')
+
         filename = os.path.join(path, rcode, 'license.lic.txt')
-        if os.path.exists(filename):
-            with open(filename) as f:
-                info = f.read()
+        if not os.path.exists(filename):
+            raise RuntimeError('No license information file')
+
+        with open(filename) as f:
+            info = f.read()
+
         return info
 
     def _get_path(self):
         c = self._config
         return os.path.join(c['homepath'], c['licpath'])
+
+    def _fetch_code(self):
+        filename = os.path.join(self._get_path(), self.counter_file)
+        if os.path.exists(filename):
+            with open(filename) as f:
+                n = int(f.read())
+        else:
+            n = 0
+        n += 1
+        with open(filename, 'w') as f:
+            f.write('%s' % n)
+        return 'reg-%06d' % n
 
 
 if __name__ == '__main__':
